+++ equipment/thermiteFlare2.lua
@@ -1,11 +1,11 @@
--- Thermite Flare (Last)
+-- Thermite Flare (2)
 
-local sprite = Resources.sprite_load("aphelion", "equipment/thermiteFlareLast", PATH.."assets/sprites/equipment/thermiteFlareLast.png", 2, 22, 22)
+local sprite = Resources.sprite_load("aphelion", "equipment/thermiteFlare2", PATH.."assets/sprites/equipment/thermiteFlare2.png", 2, 22, 22)
 
-local equip = Equipment.new("aphelion", "thermiteFlareLast", true)
+local equip = Equipment.new("aphelion", "thermiteFlare2", true)
 equip:set_sprite(sprite)
 equip:set_loot_tags(Item.LOOT_TAG.category_damage)
-equip:set_cooldown(30)
+equip:set_cooldown(6)
 
 equip:onUse(function(actor)
     local use_dir = actor:get_equipment_use_direction()
@@ -16,5 +16,5 @@ equip:onUse(function(actor)
     instData.parent = actor
     instData.hsp = 8.0 * use_dir
 
-    actor:set_equipment(-1)
+    actor:set_equipment(Equipment.find("aphelion-thermiteFlareLast"))
 end)